add's search_service need to update find a bug on goods_server

+++ ly-common/src/main/java/com/leyou/common/vo/PageResult.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class PageResult<T> {
     private Long total;// 总条数
-    private Long totalPage;// 总页数
+    private Integer totalPage;// 总页数
     private List<T> items;// 当前页数据
 
     public PageResult() {
@@ -18,7 +18,7 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public PageResult(Long total, Long totalPage, List<T> items) {
+    public PageResult(Long total, Integer totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
